cleanup firesbase config, turn trigger e2e into a mocha test, revert package-lock.json
