+++ scripts/test-triggers-end-to-end.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cd ./scripts/triggers-end-to-end-tests
+
+npm test
