--- scripts/triggers-end-to-end-tests/database.rules.json
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    ".read": false,
-    ".write": false
-  }
-}
