+++ scripts/triggers-end-to-end-tests/firebase.json
@@ -1,16 +1,9 @@
 {
-  "database": {
-    "rules": "database.rules.json"
-  },
+  "database": {},
   "firestore": {
-    "rules": "firestore.rules",
-    "indexes": "firestore.indexes.json"
+    "rules": "firestore.rules"
   },
-  "functions": {
-    "predeploy": [
-      "npm --prefix \"$RESOURCE_DIR\" run lint"
-    ]
-  },
+  "functions": {},
   "emulators": {
     "database": {
       "port": 9000
