+++ scripts/triggers-end-to-end-tests/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "triggers-end-to-end-tests",
+  "version": "1.0.0",
+  "description": "End-to-end tests for emulator cloud function triggers",
+  "main": "run.spec.js",
+  "scripts": {
+    "test": "./node_modules/mocha/bin/mocha run.spec.js --reporter spec --exit"
+  },
+  "dependencies": {
+    "grpc": "^1.21.0",
+    "mocha": "^5.0.5"
+  },
+  "author": "",
+  "license": "ISC"
+}
